improve rspc error handling

+++ apps/api-server/src/cron_jobs.rs
@@ -2,27 +2,24 @@ use content_base::{delete::DeletePayload, ContentBase};
 use content_library::Library;
 use prisma_lib::{asset_object, file_handler_task};
 use std::sync::Arc;
-use tracing::error;
 
-pub async fn delete_unlinked_assets(
-    library: Arc<Library>,
-    content_base: Arc<ContentBase>,
-) -> Result<(), rspc::Error> {
+pub async fn delete_unlinked_assets(library: Arc<Library>, content_base: Arc<ContentBase>) {
     // 查找所有未关联 file path 的 asset object
     // TODO 怎么优化更好点
-    let all_assets = library
+    let all_assets = match library
         .prisma_client()
         .asset_object()
         .find_many(vec![])
         .with(asset_object::file_paths::fetch(vec![]).take(1))
         .exec()
         .await
-        .map_err(|e| {
-            rspc::Error::new(
-                rspc::ErrorCode::InternalServerError,
-                format!("failed to find asset: {}", e),
-            )
-        })?;
+    {
+        Ok(assets) => assets,
+        Err(e) => {
+            tracing::error!("failed to find asset_object: {}", e);
+            return;
+        }
+    };
 
     let unlinked_assets = all_assets
         .into_iter()
@@ -37,42 +34,34 @@ pub async fn delete_unlinked_assets(
     tracing::debug!("unlinked_assets : {:?}", unlinked_assets);
 
     for asset in unlinked_assets {
-        library
+        if let Err(e) = library
             .prisma_client()
             .asset_object()
             .delete(asset_object::id::equals(asset.id))
             .exec()
             .await
-            .map_err(|e| {
-                rspc::Error::new(
-                    rspc::ErrorCode::InternalServerError,
-                    format!("failed to find asset: {}", e),
-                )
-            })?;
+        {
+            tracing::error!("failed to delete asset_object: {}", e);
+        }
 
-        library
+        if let Err(e) = library
             .prisma_client()
             .file_handler_task()
             .delete_many(vec![file_handler_task::asset_object_id::equals(asset.id)])
             .exec()
             .await
-            .map_err(|e| {
-                rspc::Error::new(
-                    rspc::ErrorCode::InternalServerError,
-                    format!("failed to delete asset related tasks: {}", e),
-                )
-            })?;
+        {
+            tracing::error!("failed to delete asset related tasks: {}", e);
+        }
 
         // delete from fs
         let file_path = library.file_path(&asset.hash);
         if let Err(e) = std::fs::remove_file(&file_path) {
-            error!("failed to delete file({}): {}", file_path.display(), e);
+            tracing::error!("failed to delete file({}): {}", file_path.display(), e);
         };
         let payload = DeletePayload::new(&asset.hash);
         if let Err(e) = content_base.delete(payload).await {
-            error!("failed to delete artifacts: {}", e);
+            tracing::error!("failed to delete artifacts: {}", e);
         }
     }
-
-    Ok(())
 }
